code/python_code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "unthermal"
authors = [{name = "Leonardo Bermeo", email = "lbermeoc@unal.edu.co"}]
description = "unthermal is a package designed for interacting with a physical thermal plant via IoT technology."
version = "0.1.13"
readme = "README.md"

requires-python = ">=3.7"
dependencies = [
    "iottalk-paho-mqtt",
    "control",
    "numpy",
    "matplotlib",
    "scipy",
    "ipympl",
    "mplcursors"
]
[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
where = ["datafiles"]
include = ["*.csv"]


[project.urls]
Homepage = "https://github.com/nebisman/UNThermal.git"
